Small bugs fixed

+++ src/yasim/simulator/dwgsim.py
@@ -3,6 +3,7 @@ import shutil
 from typing import List
 
 from commonutils import ioctl
+from commonutils.compressor import gz_decompress
 from yasim.simulator import Simulator
 
 
@@ -18,18 +19,43 @@ class SimulatorDwgsim(Simulator):
         return cmd
 
     def move_file_after_finish(self):
-        try:
-            shutil.move(
-                self.tmp_prefix + ".bwa.read1.fastq.gz",
-                self.output_fastq_prefix + "_1.fq.gz"
-            )
-            shutil.move(
-                self.tmp_prefix + ".bwa.read2.fastq.gz",
-                self.output_fastq_prefix + "_2.fq.gz"
-            )
-        except Exception:
-            pass
-        for filename in glob.glob(self.tmp_prefix + "*"):
+        try_read1_suffix = (
+            ".bwa.read1.fastq.gz",
+            ".bwa.read1.fastq",
+            ".bwa.read1.fq.gz",
+            ".bwa.read1.fq"
+        )
+        try_read2_suffix = (
+            ".bwa.read2.fastq.gz",
+            ".bwa.read2.fastq",
+            ".bwa.read2.fq.gz",
+            ".bwa.read2.fq"
+        )
+        for suffix_r1, suffix_r2 in zip(try_read1_suffix, try_read2_suffix):
+            if not ioctl.file_exists(self.tmp_prefix + suffix_r1) or not ioctl.file_exists(self.tmp_prefix + suffix_r2):
+                continue
+            if suffix_r1.endswith(".gz"):
+                gz_decompress(
+                    self.tmp_prefix + suffix_r1,
+                    self.output_fastq_prefix + f"_1.fq"
+                )
+                gz_decompress(
+                    self.tmp_prefix + suffix_r2,
+                    self.output_fastq_prefix + f"_2.fq"
+                )
+            else:  # Need to archive
+                shutil.move(
+                    self.tmp_prefix + suffix_r1,
+                    self.output_fastq_prefix + f"_1.fq"
+                )
+                shutil.move(
+                    self.tmp_prefix + suffix_r2,
+                    self.output_fastq_prefix + f"_2.fq"
+                )
+            break
+        else:
+            self.lh.error(f"Unable to find output")
+        for filename in glob.glob(self.tmp_prefix + ".bfast.fastq*"):
             ioctl.rm_rf(filename)
 
     def run(self) -> None:
